replaced create site page with createSite dialog
